requirements.txt
flet==0.25.2
PyYAML==6.0.1
pyinstaller==6.0.0
openpyxl==3.1.2
pandas==2.1.4
numpy==1.26.2